requirements.txt.txt
aiogram
openai
feedgen
python-dotenv
aiohttp